+++ ex2/UnionFind.h
@@ -10,17 +10,17 @@ class UFNode{
         Company* company;
         Group* group;
         //added value for companies which purchased other companies according to the requirements of the exercise
-        double fixer;
+        long double fixer;
         UFNode* parent;
     public:
         UFNode(Company* company, Group* group):company(company), group(group), fixer(0), parent(nullptr){};
         ~UFNode(){}
         Company* getCompany(){return company;}
         Group* getGroup(){return group;}
-        double getFixer(){return fixer;}
+        long double getFixer(){return fixer;}
         UFNode* getParent(){return parent;}
         void setGroup(Group* group){this->group = group;}
-        void setFixer(double fixer){this->fixer = fixer;}
+        void setFixer(long double fixer){this->fixer = fixer;}
         void setParent(UFNode* parent){this->parent = parent;}
 };
 
@@ -28,7 +28,7 @@ typedef struct Group
 {
     int group_num;
     int num_elements;
-    double total_value;
+    long double total_value;
     UFNode* root;
 }Group;
 
@@ -48,10 +48,10 @@ class UnionFind{
          * @param group_b group_b id
          * @param factor all companies in group b add factor*(group_a.value)
          */
-        void Union(int group_a, int group_b, double factor);
+        void Union(int group_a, int group_b, long double factor);
         //returns company name
         int Find(int company_id);
-        double getCompanyValue(int company_id);
+        long double getCompanyValue(int company_id);
         //O(1)
         Company* getCompany(int company_id);
 };
